requirements.txt
ctransformers
sentence-transformers
langchain-huggingface
langchain
langchain-community
langchain-pinecone
flask
pypdf
python-dotenv
-e .